Remove StarknetSequencerTransaction; Use StarknetTransaction instead
