+++ Sources/Starknet/Data/Transaction/TransactionProtocol.swift
@@ -1,9 +1,6 @@
 
-public protocol StarknetSequencerTransaction: Codable, Hashable, Equatable {
+public protocol StarknetTransaction: Codable, Hashable, Equatable {
     var type: StarknetTransactionType { get }
     var version: Felt { get }
-}
-
-public protocol StarknetTransaction: StarknetSequencerTransaction {
     var hash: Felt? { get }
 }
